requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
PyQt6==6.7.1
python-dateutil==2.9.0.post0